% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{catchbysau}
\alias{catchbysau}
\title{catchbysau calculates the catch by sau with error for Tasmania}
\usage{
catchbysau(inexpB, sauindex, TAC, sigmab)
}
\arguments{
\item{inexpB}{a particular years' exploitable biomass by population}

\item{sauindex}{a vector containing the sau index number for each population}

\item{TAC}{the sum of the aspirational catches for a given year}

\item{sigmab}{the sd of the log-normal errors included in the estimates of
the exploitable biomass by SAU}
}
\value{
a vector of the catches by sau
}
\description{
catchbysau calculates the catch by sau with error for the
Tasmanian HS. This takes the sum of the aspirational catches as the TAC,
which is multiplied by the proportion of exploitable biomass for each SAU,
which has log-Normal error included with sd=sigmaB.
}
\examples{
print("wait on suitable internal data")
}
