% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tas_specific.R
\name{checkhsargs}
\alias{checkhsargs}
\title{checkhsargs review the hsargs and sends a summary to the console}
\usage{
checkhsargs(x)
}
\arguments{
\item{x}{the hsargs object}
}
\value{
nothing
}
\description{
checkhsargs reviews the arguments within hsargs and summarizes
them to the console. Now that hsargs is getting relatievly complex,
mistakes are easier to make so this provides a review to ensure each
scenario is being run as desired.
}
\examples{
print("wait on data")
}
