% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constC.R
\name{constdata}
\alias{constdata}
\title{constdata processes the projections to produce the data for the HCR}
\usage{
constdata(
  tasCPUE,
  tasFIS,
  tasNaS,
  sauCPUE,
  sauacatch,
  saucatch,
  sauNAS,
  year,
  decrement = NULL,
  startCE = NULL
)
}
\arguments{
\item{tasCPUE}{a function that generates the CPUE statistics}

\item{tasFIS}{a function that generates the FIS statistics}

\item{tasNaS}{a function that generates the Numbers-at-size samples}

\item{sauCPUE}{the cpue by sau object from the dynamics object from the
replicate model runs}

\item{sauacatch}{aspirational catches for each SAU from year-1}

\item{saucatch}{actual catches for each SAU from year-1, these are not needed
in Tasmania or Victoria, but are needed (as well as acatch) in South
Australia}

\item{sauNAS}{the numbers-at-size data from the zoneDP object}

\item{year}{the specific year being considered}

\item{decrement}{merely present to allow for HCR other than consthcr, hence
set to NULL}

\item{startCE}{merely present to allow for HCR other than consthcr, hence
set to NULL}
}
\value{
a list of arrce, yearnames, acatch, fis, and nas
}
\description{
constdata takes the projection results from zoneDP, and histCE
from 'otherdata', and appropriately samples them ready for the HCR
}
\examples{
print("wait on suitable data sets")
# constCPUE=sampleCE;constFIS=sampleFIS;constNAS=sampleNaS;sauCPUE=zoneDP$cesau;year=year;iter=iter;
# sauacatch=zoneDP$acatch
}
