% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constC.R
\name{consthcr}
\alias{consthcr}
\title{consthcr conducts the MCDA and returns the TAC multiplier and details}
\usage{
consthcr(indat, hsargs, glb, projyear = NULL, outhcr = outhcr, iter = iter)
}
\arguments{
\item{indat}{the output from constdata = NULL, just a dummy variable.}

\item{hsargs}{a vector setting aspirational catches by SAU. They should sum
to the TAC}

\item{glb}{the globals object}

\item{projyear}{an index of the year for which the acatch are being estimated
thus, if one has 29 years of data hen projyear would be 30.}

\item{outhcr}{a list of arrays generated by makehcrout within the HS package.
It becomes the repository of all the HS scores and other outputs.}

\item{iter}{which iteration within reps is being run? Needed to output the
scores to outhcr}
}
\value{
a list of the acatch and TAC
}
\description{
consthcr enables one to implement constant aspirational catches
for each sau, this can act as a form of status quo approach. All it
requires is for the hsargs = a vector of nsau aspirational catches. Then,
in the do-MSE function replace 'hcrfun=mcdahcr' with 'hcrfun=consthcr',
'getdata=tasdata' with 'getdata=constdata, and finally,
'makeouthcr=makeouthcr' with 'makeouthcr=makeoutconst'
}
\examples{
print("wait on data and time")
# indat=constdata; hsargs=hsargs;
}
