% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{getcpueHS}
\alias{getcpueHS}
\title{getcpueHS recreates all the TasHS outputs from the MSE projections}
\usage{
getcpueHS(ce, catches, glb, yearCE, hsargs, sau)
}
\arguments{
\item{ce}{the (hyrs+pyrs) x nsau x reps array of cpue for the scenario found
in sauout$zonePsau$cpue}

\item{catches}{the (hyrs+pyrs) x nsau x reps array of actual catches for the
scenario found in sauout$zonePsau$cpue}

\item{glb}{the globals object}

\item{yearCE}{the year names for the observed CPUE used to condition the
model, found in condC$yearCE. Used to identify the observed and projected
years in the simulations.}

\item{hsargs}{the input arguments used for the TasHS in the given scenario}

\item{sau}{which sau should the analysis be conducted upon}
}
\value{
a list containing the cpue and catch series used, then, for each year, the
cetarget, the grad1score, the grad4score, the target score, and the
final weighted score
}
\description{
getcpueHS is designed to take the cpue from the finished
projections and then for a given sau, it calculates for each year of the
projections the target cpue, the grad1score, grad4score, and targscore,
as well as the final combined score. In addition, the cpue time series
used in the calculations is also output. Used directly by aMSE.
}
\examples{
print("wait on data sets")
# ce=cpue;catches=catches;glb=glb;yearCE=yearCE;hsargs=hsargs;sau=sau
}
