% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{getgrad1}
\alias{getgrad1}
\title{getgrad1 calculates the one year gradient score for all years of cpue}
\usage{
getgrad1(vectce)
}
\arguments{
\item{vectce}{vector of cpue for a given spatial scale}
}
\value{
a vector of gradient1 for each year, starting from year 2
}
\description{
getgrad1 calculates the one year gradient score for a
vector of CPUE. The equation used is (CE(2:y) / CE(1:y-1)) - 1,
which provides the annual proportional change in CPUE.
}
\examples{
  data(hcrdata)
  cpue <- hcrdata$arrce
  nyr <- nrow(cpue)
  grad1 <- getgrad1(cpue[,1])
  score1 <- getscore(grad1)
  cbind(cpue[,1],grad1,score1)
}
