% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{getgrad4}
\alias{getgrad4}
\title{getgrad4 applies a linear regression in steps of wid to input}
\usage{
getgrad4(vectce, wid = 4)
}
\arguments{
\item{vectce}{the input vector of cpue}

\item{wid}{the number of years of the input data to use in each
regression}
}
\value{
a vector of length (wid-1) shorter than the input vector
}
\description{
getgrad4 takes an input vector of cpue and, in chunks
of length wid, converts them to proportional changes by
dividing through by the first value of the short series, then
applies a linear regression keeping only the gradient.
}
\examples{
  data(hcrdata)
  cpue <- hcrdata$arrce
  nyr <- nrow(cpue)
  x <- cpue[,1]
  grad4 <- getgrad4(x,wid=4)
  grad3 <- getgrad4(x,wid=3)
  cbind(x,grad4,grad3)
}
