% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{getscore}
\alias{getscore}
\title{getscore calculates the scores for the grad1 and grad4 PMs}
\usage{
getscore(pm, mult = 0.1)
}
\arguments{
\item{pm}{the raw performance measure values derived from the functions
getgrad1,  getgrad4, or targscore}

\item{mult}{the multiplier on the bounds to expand them upwards and
downwards. default value = 1.1 = 10 percent increase; from hsargs$mult}
}
\value{
a vector of scores to be included in the MCDA
}
\description{
getscore calculates the scores for the grad1 and grad4
performance measures. It does this by re-scaling the range of
the PM values and then fitting separate linear regressions to
the values above and below zero. These enable it to calculate
the predicted scores.
}
\examples{
  data(hcrdata)
  cpue <- hcrdata$arrce
  nyr <- nrow(cpue)
  grad1 <- getgrad1(cpue[,1])
  score1 <- getscore(grad1)
  cbind(cpue[,1],grad1,score1)
}
\seealso{
{
\link{getgrad1}, \link{getgrad4}, \link{targscore}
}
}
