% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constC.R
\name{makeoutconst}
\alias{makeoutconst}
\title{makeouthcr creates an object to contain the replicate HCR outputs}
\usage{
makeoutconst(glb, hsargs)
}
\arguments{
\item{glb}{the globals object for the MSE}

\item{hsargs}{the harvest strategy arguments, in this case a vector of
aspirational catches for each sau}
}
\value{
in this case it merely returns the vector of catches in hsargs
}
\description{
makeouthcr creates an objects required to hold the cumulative
values for the grad4 pm. Otherwise the full run from 1992 - endyr must be
recalculated each time. In this case it is a 3-D array
set up to hold the values used on each SAUs previous years
aspirational catch. Copies of the output array could also be used to
store the scores developed for the different components of the HCR.
}
\examples{
makeoutconst(glb=NULL,hsargs=c(10,20,30,40,50,60,70,80))
}
