% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{makeouthcr}
\alias{makeouthcr}
\title{makeouthcr creates an object to contain the replicate HCR outputs}
\usage{
makeouthcr(glb, hsargs)
}
\arguments{
\item{glb}{the globals object for the MSE}

\item{hsargs}{the input list of arguments used by the HCR}
}
\value{
in this case a list of g1s, g4s, targsc, finalsc, index, catchmult,
metaflag, and cetarg
}
\description{
makeouthcr creates an objects required to hold the cumulative
values for the grad4 pm. Otherwise the full run from 1992 - endyr must be
recalculated each time. In this case it is a 3-D array
set up to hold the values used on each SAUs previous years
aspirational catch. Copies of the output array could also be used to
store the scores developed for the different components of the HCR.
Currently it stores 3-D arrays of years x sau x reps for g1s = gradient 1
score, g4s = gradient 4 score, targsc = target CE score, finalsc = final
weighted score (which can be modified by metarule 3), index = the index
of the vector of hcr aspirational catch multipliers, catchmult = the
value from the hcr vector in hsargs (can be modified by metarule 1 and 2),
metaflag = a flag for when a metarule is triggered where meta rule 3 = 3,
rule 2 = 2, and meta rule 1 = 1.25 (so when combined the toals can be
identifed), and cetarg
}
\examples{
args(makeouthcr)
}
