% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{ploteHSphase}
\alias{ploteHSphase}
\title{ploteHSphase generates a phase plot of CPUE Target vs Gradient scores}
\usage{
ploteHSphase(cpue, rundir = "", wid = 4, fnt = 7, cex = 1, console = TRUE)
}
\arguments{
\item{cpue}{a vector of cpue data either predicted or observed.}

\item{rundir}{the directory for all files and results, default=""}

\item{wid}{default=4 the number of years used on getgrad4 from TasHS}

\item{fnt}{default = 7 = times bold.}

\item{cex}{default font size = 1.0}

\item{console}{should the plot go to the console or be saved? Default=TRUE}
}
\value{
nothing but it does plot a graph
}
\description{
The empirical Tasmanian harvest strategy uses different CPUE
statistics to produce scores that represent different facets of CPUE.
Plotting the score for the CPUE gradient measure against the CPUE target
measure is used as a proxy for a phase plot of fishing mortality against
mature biomass. Strictly, statistics relating to CPUE are concerned with
the exploitable biomass, which, because of the LML, cam be markedly
different from the spawning biomass. Nevertheless, trend in one tend to
be similar to trends in the other.
}
\examples{
print("wait on a data set")
}
