% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tas_specific.R
\name{plotfinalscores}
\alias{plotfinalscores}
\title{plotfinalscores plots projected catches, cpue, PMs, and final score}
\usage{
plotfinalscores(
  outhcr,
  zoneDP,
  sau,
  minprojC = 0,
  minprojCE = 0,
  mintargCE = 0,
  filen = ""
)
}
\arguments{
\item{outhcr}{the HS scores objects generates by mcdahcr or constantrefhcr}

\item{zoneDP}{the object containing the dynamics of the projected zone}

\item{sau}{which sau index is to have its outputs plotted. With the 8 sau in
the western zone this index is 1:8, where 1 = 6W, 2= sau7, 3=sau8, etc.}

\item{minprojC}{the minimum y-axis value for the projected catches}

\item{minprojCE}{the minimum y-axis value for the projected cpue}

\item{mintargCE}{the minimum y-axis value for the projected target cpue}

\item{filen}{default="", this would be set by the process running the
function for each sau}
}
\value{
invisibly a matrix of the median catches, cpue, targetce, finalscore,
grad1 score, grad4 score, and target cpue score.
}
\description{
plotfinalscores takes the individual sau output from getcpueHS
and generates a plot made up of six sub-plots of the replicates for each
sau. In sequence these plots are the projected catches, the projected
cpue, the gradient 1 scores, the gradient 4 scores, the target cpue, and
the final total score. This enables the relationships between the scores
to be examined. Only used in do_MSE
}
\examples{
print("wait on data sets")
}
