% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tas_specific.R
\name{plotmultandflags}
\alias{plotmultandflags}
\title{plotmultandflags plots projected catches, cpue, catchmult, and meta-flags}
\usage{
plotmultandflags(outhcr, sauans, sau, minprojC = 0, minprojCE = 0, filen = "")
}
\arguments{
\item{outhcr}{the HS scores objects generates by mcdahcr or constantrefhcr}

\item{sauans}{the object containing the dynamics of the projected zone
compressed to the sau level, this is sauout, which is dervied from the
sauplots function}

\item{sau}{which sau index is to have its outputs plotted. With the 8 sau in
the western zone this index is 1:8, where 1 = 6W, 2= sau7, 3=sau8, etc.}

\item{minprojC}{the minimum y-axis value for the projected catches}

\item{minprojCE}{the minimum y-axis value for the projected cpue}

\item{filen}{default = "", which implies the plot goes to the console.}
}
\value{
invisibly a list of the median catches, cpue, catchmult, and
meta-flag
}
\description{
plotmultandflags takes the individual sau output from outhcr
and generates a plot made up of four sub-plots of the replicates for each
sau. In sequence these plots are the projected catches, the projected
cpue, the catchmult values, and a histogram of the metaflag values if
there are any meta flags raised. If no flags then a null plot is made.
Only used in do_MSE
}
\examples{
print("wait on data sets")
# outhcr=outhcr;sauans=sauout;sau=1;minprojC=0;minprojCE=0;filen=""
}
