% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{targscore}
\alias{targscore}
\title{targscore generates the HCR score for the target PM}
\usage{
targscore(vectce, qnt = 0.55, mult = 0.1, maxtarg = 150)
}
\arguments{
\item{vectce}{the vector of cpue to be assessed}

\item{qnt}{the quantile of the input vector selected as the target,
default = 0.55}

\item{mult}{the multiplier on the bounds to expand them upwards and
downwards. default value = 1.1 = 10 percent increase and decrease}

\item{maxtarg}{a meta-rule that sets an upper limit on the target cpue, the
default=150kg/hr}
}
\value{
a list of the final year's score, the internals to the calculations,
and he target and limit reference points
}
\description{
targscore takes in a vector of cpue x years and defines the targ
from the complete series, and the limit reference point. This differs
from the two getgradone and getgradwid in needing multiple years at once.
One meta-rule is that the target cpue should not rise above
}
\examples{
print("wait on suitable data")
args(targscore)
# vectce=40:140; qnt=0.55;mult=0.1; maxtarg=150.0
}
