% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{targscoreconstref}
\alias{targscoreconstref}
\title{targscoreconstref generates the HCR score for the fixed target PM}
\usage{
targscoreconstref(vectce, realtarg, mult = 0.1, maxtarg = 150)
}
\arguments{
\item{vectce}{the vector of cpue to be assessed}

\item{realtarg}{the quantile of the reference period of cpue x years used as
the target. The default quantile = 0.55}

\item{mult}{the multiplier on the bounds to expand them upwards and
downwards. default value = 1.1 = 10 percent increase and decrease}

\item{maxtarg}{a meta-rule that sets an upper limit on the target cpue, the
default=150kg/hr}
}
\value{
a list of the final year's score, the internals to the calculations,
and he target and limit reference points
}
\description{
targscoreconstref takes in a vector of cpue x years, where the
the years used are defined in hsargs as startCE and endCE, which for
Tasmania default to 1992:2012 and defines the targ for each sau.
}
\examples{
print("wait on suitable data")
args(targscore)
# vectce=40:140; qnt=0.55;mult=0.1; maxtarg=150.0
}
