% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{tasCatch}
\alias{tasCatch}
\title{tasCatch extracts the catch data if required by the associated HS}
\usage{
tasCatch(x)
}
\arguments{
\item{x}{a dummy variable not used}
}
\description{
tasCatch no catch data are currently required by the
Tasmanian HCR so this function merely returns NULL. Used directly by
aMSE.
}
\examples{
x <- tasCatch(NA)
print(x)
}
