% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{tasCPUE}
\alias{tasCPUE}
\title{tasCPUE prepares the projected cpue data ready for the tasHCR}
\usage{
tasCPUE(cesau, year, startCE, decrement = 1)
}
\arguments{
\item{cesau}{the projected cpue by SAU for a particular year and iteration
from zoneDP}

\item{year}{the year of projection is used, along with 'startCE' to define
which years of cpue data to use.}

\item{startCE}{the first year of observed cpue data to use. In Tasmanian
nothing earlier than 1992 is considered sufficiently reliable for use.}

\item{decrement}{how many years from the current to remove from the cpue
cpue time series, default=1}
}
\value{
a list of the array of cpue and the yearnames vector
}
\description{
tasCPUE prepares the projected cpue data and the yearnames from
zoneDP ready for their use within the tasHCR. The yearnames identifies
the extent of the years of data being included in the calculation of the
cpue performance measures. In the MSE (not in reality), to ensure that
predicted values for each performance measure are available, the first
year of the predicted cpue data to be used has to be two years prior to
the first year of useful observed cpue. This is particularly necessary so
that the grad4 performance measure has four years of data to work with.
Used directly by aMSE.
}
\examples{
args(tasCPUE)
# tasCPUE(cesau=sauCPUE,year=year,startCE=hsargs$startCE,decrement=1)
# cesau=zoneDP$cesau; iter=1; year=59
# cesau=sauCPUE; year=62; startCE=hsargs$startCE; decrement=1
}
