% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{tasFIS}
\alias{tasFIS}
\title{tasFIS calculates required FIS data}
\usage{
tasFIS(x)
}
\arguments{
\item{x}{a dummy variable not used}
}
\description{
tasFIS no FIS data are currently required by the Tasmanian
HCR so this function merely returns NULL. Used directly by aMSE.
}
\examples{
x <- tasFIS(NA)
print(x)
}
