% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashs_funs.R
\name{tasNaS}
\alias{tasNaS}
\title{tasNas calculates required Numbers-at-Size data}
\usage{
tasNaS(x)
}
\arguments{
\item{x}{a dummy variable not used}
}
\description{
tasNas no numbers-at-size data are currently required by the
Tasmanian HCR so this function merely returns NULL. Used directly by
aMSE.
}
\examples{
x <- tasNaS(NA)
print(x)
}
